feat: separate read and write database contexts, update references across services and repositories #67
