+++ Services/Order/src/Order.Api/Program.cs
@@ -10,7 +10,8 @@ using Order.Application.Commands.CreateOrder;
 using Order.Application.Integrations;
 using Order.Application.Interfaces;
 using Order.Domain.Interfaces;
-using Order.Infra.Data.Context;
+using Order.Infra.Data.Context.Read;
+using Order.Infra.Data.Context.Write;
 using Order.Infra.Data.Mongo;
 using Order.Infra.Data.Repositories;
 using Order.Infra.Data.Services;
