+++ Services/Order/src/Order.Infra/Data/Context/Write/WriteDbContext.cs
@@ -1,7 +1,7 @@
 using BuildingBlocks.SharedKernel.Config;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.Extensions.Options;
-namespace Order.Infra.Data.Context;
+namespace Order.Infra.Data.Context.Write;
 
 public class WriteDbContext : DbContext
 {
