+++ Services/Order/src/Order.Infra/Data/Mongo/MongoDbInitializer.cs
@@ -1,6 +1,6 @@
 ﻿using BuildingBlocks.Infra.MongoReadModels;
 using MongoDB.Driver;
-using Order.Infra.Data.Context;
+using Order.Infra.Data.Context.Read;
 using Serilog;
 
 namespace Order.Infra.Data.Mongo;
